refactor(ecsi): inline expanded dynamics and simplify alignment
